+++ Programming/EcoBike/src/entity/bike/TwinElectricBike.java
@@ -24,7 +24,7 @@ public class TwinElectricBike extends StandardElectricBike {
         bike.setNumRearSeat(res.getInt("numRearSeat"));
         bike.setLicensePlate(res.getString("licensePlate"));
         bike.setNumPedal(res.getInt("numPedal"));
-        bike.setValue(res.getDouble("value"));
+        bike.setValue(res.getInt("value"));
         bike.setCoefficient(res.getInt("coefficientPrice"));
         bike.setUrlImage(res.getString("urlImage"));
         bike.setNumSaddle(res.getInt("numSaddle"));
